Moved some files from app folder to components folder to show that they are used in multiple pages now
